moving-address-fields-to-common-component-and-adding-address-to-department-page
fixed "user","department" and "employee" folder name to lowercase
