+++ EmployeePortal/ClientApp/src/App.js
@@ -6,16 +6,16 @@ import ComponentWrapper from './components/helpers/ComponentWrapper';
 import { AuthRoutes } from './components/helpers/RequireAuth';
 import { Errors } from './components/Pages/Admin/Errors';
 import { Logs } from './components/Pages/Admin/Logs';
-import { ChangePassword } from './components/Pages/Admin/User/ChangePassword';
-import { CreateUser } from './components/Pages/Admin/User/CreateUser';
-import { EditUserProfile } from './components/Pages/Admin/User/EditUserProfile';
-import { UserProfile } from './components/Pages/Admin/User/UserProfile';
-import { Users } from './components/Pages/Admin/User/Users';
-import { AddDepartment } from './components/Pages/Department/AddDepartment';
-import { Departments } from './components/Pages/Department/Departments';
-import { EditDepartment } from './components/Pages/Department/EditDepartment';
-import { AddEditEmployee } from './components/Pages/Employee/AddEditEmployee';
-import { Employees } from './components/Pages/Employee/Employees';
+import { ChangePassword } from './components/Pages/Admin/user/ChangePassword';
+import { CreateUser } from './components/Pages/Admin/user/CreateUser';
+import { EditUserProfile } from './components/Pages/Admin/user/EditUserProfile';
+import { UserProfile } from './components/Pages/Admin/user/UserProfile';
+import { Users } from './components/Pages/Admin/user/Users';
+import { AddDepartment } from './components/Pages/department/AddDepartment';
+import { Departments } from './components/Pages/department/Departments';
+import { EditDepartment } from './components/Pages/department/EditDepartment';
+import { AddEditEmployee } from './components/Pages/employee/AddEditEmployee';
+import { Employees } from './components/Pages/employee/Employees';
 import { ErrorPage } from './components/Pages/ErrorPage';
 import { Home } from './components/Pages/Home';
 import { NotFoundPage } from './components/Pages/NotFoundPage';
